+++ packages/aimd-recorder/vitest.config.ts
@@ -1,8 +1,9 @@
+import vue from '@vitejs/plugin-vue'
 import { defineConfig } from 'vitest/config'
 import tsconfigPaths from 'vite-tsconfig-paths'
 
 export default defineConfig({
-  plugins: [tsconfigPaths()],
+  plugins: [vue(), tsconfigPaths()],
   test: {
     include: ['src/**/*.test.ts'],
     environment: 'happy-dom',